third_party/blink/web_tests/virtual/text-antialias/custom-font-data-crash-expected.txt

CONSOLE WARNING: Failed to decode downloaded font: custom-font-data-crash.html
CONSOLE WARNING: OTS parsing error: invalid version tag
PASS
